Remove all axum-sessions-auth stuff.
Was finding it seemed to take too much building of what's already
managed by the API to use this, and yet still couldn't simply get a
User's info from the session (the other main benefit to me). Instead
will be falling back on a "simpler" login server fn that defers to the
API for token creation, then stuffs it in an http only cookie and sets
that with a response header.

Additionally will defer to the API to check token validity and also get
User info from a token. May end up building logic for refreshing tokens
into the App and API at some point, but that's future me's problem.